+++ pkg/plugin/internal/handlers/query_data.go
@@ -14,6 +14,8 @@ import (
 	"github.com/hashicorp/go-hclog"
 	"gopkg.in/yaml.v3"
 
+	"github.com/Enapter/telemetry-grafana-datasource-plugin/pkg/commandsapi"
+	"github.com/Enapter/telemetry-grafana-datasource-plugin/pkg/httperr"
 	"github.com/Enapter/telemetry-grafana-datasource-plugin/pkg/telemetryapi"
 )
 
@@ -22,12 +24,17 @@ var _ backend.QueryDataHandler = (*QueryData)(nil)
 type QueryData struct {
 	logger             hclog.Logger
 	telemetryAPIClient telemetryapi.Client
+	commandsAPIClient  commandsapi.Client
 }
 
-func NewQueryData(logger hclog.Logger, telemetryAPIClient telemetryapi.Client) *QueryData {
+func NewQueryData(
+	logger hclog.Logger, telemetryAPIClient telemetryapi.Client,
+	commandsAPIClient commandsapi.Client,
+) *QueryData {
 	return &QueryData{
 		logger:             logger.Named("query_handler"),
 		telemetryAPIClient: telemetryAPIClient,
+		commandsAPIClient:  commandsAPIClient,
 	}
 }
 
@@ -52,115 +59,24 @@ func (h *QueryData) QueryData(
 		}
 	}
 
-	h.makeLabelsUnique(resp.Responses)
-
 	return resp, nil
 }
 
-//nolint:gocognit // FIXME
-func (h *QueryData) makeLabelsUnique(responses backend.Responses) {
-	defaultNames := make(map[string]string, len(responses))
-	frames := make(map[string]*data.Frame)
-	numFields := 0
-	const oneForTimeField = 1
-
-	type kvpair struct {
-		k string
-		v string
-	}
-
-	counter := make(map[kvpair]int)
-
-	for refID, resp := range responses {
-		if resp.Error != nil {
-			continue
-		}
-
-		switch {
-		case len(resp.Frames) == 0:
-			continue
-		case len(resp.Frames) == 1:
-			// ok
-		case len(resp.Frames) > 1:
-			h.logger.Warn("multiple data frames are not supported: " +
-				"skip making labels unique")
-			return
-		default:
-			panic("unreachable")
-		}
-
-		frame := resp.Frames[0]
-		if len(frame.Fields) < oneForTimeField+1 {
-			continue
-		}
-
-		frames[refID] = frame
-
-		for _, field := range frame.Fields[oneForTimeField:] {
-			for k, v := range field.Labels {
-				if k == "telemetry" {
-					defaultNames[refID] = v
-				}
-				counter[kvpair{k, v}]++
-			}
-			numFields++
-		}
-	}
-
-	for kv, n := range counter {
-		if n == numFields {
-			for _, frame := range frames {
-				for _, field := range frame.Fields[oneForTimeField:] {
-					delete(field.Labels, kv.k)
-				}
-			}
-		}
-	}
-
-	for refID, frame := range frames {
-		for _, field := range frame.Fields[oneForTimeField:] {
-			if len(field.Labels) == 0 {
-				field.Name = defaultNames[refID]
-			}
-		}
-	}
-}
-
-func (h *QueryData) userFacingError(err error) error {
-	if errors.Is(err, errUnsupportedTimeseriesDataType) {
-		return ErrMetricDataTypeIsNotSupported
-	}
-
-	if e := (&yaml.TypeError{}); errors.As(err, &e) {
-		return ErrInvalidYAML
-	}
-
-	var multiErr *telemetryapi.MultiError
-
-	if ok := errors.As(err, &multiErr); !ok {
-		return ErrSomethingWentWrong
-	}
-
-	switch len(multiErr.Errors) {
-	case 0: // should never happen
-		h.logger.Error("multi error does not contains errors")
-		return ErrSomethingWentWrong
-	case 1:
-		// noop
-	default:
-		h.logger.Warn("multi error contains multiple errors, " +
-			"but this is not supported yet; will return only the first error")
-	}
-
-	if msg := multiErr.Errors[0].Message; len(msg) > 0 {
-		//nolint: goerr113 // user-facing
-		return errors.New(msg)
-	}
-
-	return ErrSomethingWentWrong
-}
-
 func (h *QueryData) handleQuery(
+	ctx context.Context, pCtx backend.PluginContext, query backend.DataQuery,
+) (data.Frames, error) {
+	if query.QueryType == "command" {
+		frames, err := h.handleCommandQuery(ctx, pCtx, query)
+		if err != nil {
+			return nil, fmt.Errorf("command: %w", err)
+		}
+		return frames, nil
+	}
+
+	return h.handleTelemetryQuery(ctx, pCtx, query)
+}
+
+func (h *QueryData) handleCommandQuery(
 	ctx context.Context, pCtx backend.PluginContext, query backend.DataQuery,
 ) (data.Frames, error) {
 	user := ""
@@ -168,7 +84,52 @@ func (h *QueryData) handleQuery(
 		user = pCtx.User.Email
 	}
 
-	var props queryProperties
+	//nolint:tagliatelle // js
+	var props struct {
+		Payload struct {
+			CommandName string                 `json:"commandName"`
+			CommandArgs map[string]interface{} `json:"commandArgs"`
+			DeviceID    string                 `json:"deviceId"`
+			HardwareID  string                 `json:"hardwareId"`
+		} `json:"payload"`
+	}
+	if err := json.Unmarshal(query.JSON, &props); err != nil {
+		return nil, fmt.Errorf("parse query properties: %w", err)
+	}
+
+	cmdResp, err := h.commandsAPIClient.Execute(ctx, commandsapi.ExecuteParams{
+		User: user,
+		Request: commandsapi.CommandRequest{
+			CommandName: props.Payload.CommandName,
+			CommandArgs: props.Payload.CommandArgs,
+			DeviceID:    props.Payload.DeviceID,
+			HardwareID:  props.Payload.HardwareID,
+		},
+	})
+	if err != nil {
+		return nil, fmt.Errorf("commands api: %w", err)
+	}
+
+	frame, err := h.commandResponseToDataFrame(cmdResp)
+	if err != nil {
+		return nil, fmt.Errorf("convert cmd resp to data frame: %w", err)
+	}
+
+	return data.Frames{frame}, nil
+}
+
+func (h *QueryData) handleTelemetryQuery(
+	ctx context.Context, pCtx backend.PluginContext, query backend.DataQuery,
+) (data.Frames, error) {
+	user := ""
+	if pCtx.User != nil {
+		user = pCtx.User.Email
+	}
+
+	var props struct {
+		Hide bool   `json:"hide"`
+		Text string `json:"text"`
+	}
 	if err := json.Unmarshal(query.JSON, &props); err != nil {
 		return nil, fmt.Errorf("parse query properties: %w", err)
 	}
@@ -198,7 +159,43 @@ func (h *QueryData) handleQuery(
 		return nil, fmt.Errorf("convert timeseries to data frame: %w", err)
 	}
 
+	h.makeLabelsUnique(frame)
+
 	return data.Frames{frame}, nil
+}
+
+func (h *QueryData) userFacingError(err error) error {
+	if errors.Is(err, errUnsupportedTimeseriesDataType) {
+		return ErrMetricDataTypeIsNotSupported
+	}
+
+	if e := (&yaml.TypeError{}); errors.As(err, &e) {
+		return ErrInvalidYAML
+	}
+
+	var multiErr *httperr.MultiError
+
+	if ok := errors.As(err, &multiErr); !ok {
+		return ErrSomethingWentWrong
+	}
+
+	switch len(multiErr.Errors) {
+	case 0: // should never happen
+		h.logger.Error("multi error does not contains errors")
+		return ErrSomethingWentWrong
+	case 1:
+		// noop
+	default:
+		h.logger.Warn("multi error contains multiple errors, " +
+			"but this is not supported yet; will return only the first error")
+	}
+
+	if msg := multiErr.Errors[0].Message; len(msg) > 0 {
+		//nolint: goerr113 // user-facing
+		return errors.New(msg)
+	}
+
+	return ErrSomethingWentWrong
 }
 
 func (h *QueryData) prepareQueryText(
@@ -261,6 +258,20 @@ func (h *QueryData) DefaultGranularity(interval time.Duration) time.Duration {
 	return granularities[len(granularities)-1]
 }
 
+func (h *QueryData) commandResponseToDataFrame(
+	resp commandsapi.CommandResponse,
+) (*data.Frame, error) {
+	payloadBytes, err := json.Marshal(resp.Payload)
+	if err != nil {
+		return nil, err
+	}
+
+	return &data.Frame{Fields: data.Fields{
+		data.NewField("state", nil, []string{resp.State}),
+		data.NewField("payload", nil, []json.RawMessage{payloadBytes}),
+	}}, nil
+}
+
 func (h *QueryData) timeseriesToDataFrame(timeseries *telemetryapi.Timeseries) (*data.Frame, error) {
 	frameFields := make([]*data.Field, len(timeseries.DataFields)+1)
 
@@ -304,4 +315,47 @@ func (h *QueryData) timeseriesToDataFrame(timeseries *telemetryapi.Timeseries) (
 	}
 
 	return data.NewFrame("", frameFields...), nil
+}
+
+func (h *QueryData) makeLabelsUnique(frame *data.Frame) {
+	if len(frame.Fields) == 0 {
+		return
+	}
+
+	const oneForTimeField = 1
+	dataFields := frame.Fields[oneForTimeField:]
+
+	const metricLabelName = "telemetry"
+	var defaultName string
+
+	type kvpair struct {
+		k string
+		v string
+	}
+	counter := make(map[kvpair]int)
+
+	for _, field := range dataFields {
+		for k, v := range field.Labels {
+			if k == metricLabelName {
+				defaultName = v
+			}
+			counter[kvpair{k, v}]++
+		}
+	}
+
+	for kv, n := range counter {
+		if n == len(dataFields) {
+			for _, field := range dataFields {
+				delete(field.Labels, kv.k)
+			}
+		}
+	}
+
+	for _, field := range dataFields {
+		if len(field.Labels) == 0 && defaultName != "" {
+			field.Labels = data.Labels{
+				metricLabelName: defaultName,
+			}
+		}
+	}
 }