+++ pkg/plugin/internal/handlers/query_properties.go
@@ -1,6 +1,0 @@
-package handlers
-
-type queryProperties struct {
-	Hide bool   `json:"hide"`
-	Text string `json:"text"`
-}